Use RestrictedAccessDelegate to handle user access
