+++ milo-examples/server-examples/src/main/java/org/eclipse/milo/examples/server/ExampleNamespace.java
@@ -25,18 +25,15 @@ import org.eclipse.milo.opcua.sdk.core.AccessLevel;
 import org.eclipse.milo.opcua.sdk.core.Reference;
 import org.eclipse.milo.opcua.sdk.core.ValueRank;
 import org.eclipse.milo.opcua.sdk.server.OpcUaServer;
-import org.eclipse.milo.opcua.sdk.server.Session;
 import org.eclipse.milo.opcua.sdk.server.api.DataItem;
 import org.eclipse.milo.opcua.sdk.server.api.MethodInvocationHandler;
 import org.eclipse.milo.opcua.sdk.server.api.MonitoredItem;
 import org.eclipse.milo.opcua.sdk.server.api.Namespace;
-import org.eclipse.milo.opcua.sdk.server.api.nodes.VariableNode;
 import org.eclipse.milo.opcua.sdk.server.nodes.AttributeContext;
 import org.eclipse.milo.opcua.sdk.server.nodes.UaFolderNode;
 import org.eclipse.milo.opcua.sdk.server.nodes.UaMethodNode;
 import org.eclipse.milo.opcua.sdk.server.nodes.UaNode;
 import org.eclipse.milo.opcua.sdk.server.nodes.UaVariableNode;
-import org.eclipse.milo.opcua.sdk.server.nodes.delegates.VariableNodeDelegate;
 import org.eclipse.milo.opcua.sdk.server.util.AnnotationBasedInvocationHandler;
 import org.eclipse.milo.opcua.sdk.server.util.SubscriptionModel;
 import org.eclipse.milo.opcua.stack.core.AttributeId;
@@ -52,7 +49,6 @@ import org.eclipse.milo.opcua.stack.core.types.builtin.QualifiedName;
 import org.eclipse.milo.opcua.stack.core.types.builtin.StatusCode;
 import org.eclipse.milo.opcua.stack.core.types.builtin.Variant;
 import org.eclipse.milo.opcua.stack.core.types.builtin.XmlElement;
-import org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.UByte;
 import org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.UInteger;
 import org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.UShort;
 import org.eclipse.milo.opcua.stack.core.types.enumerated.NodeClass;
@@ -275,62 +271,13 @@ public class ExampleNamespace implements Namespace {
 
         node.setValue(new DataValue(new Variant("shh... don't tell the lusers")));
 
-        node.setAttributeDelegate(new VariableNodeDelegate() {
-            @Override
-            protected UByte getUserAccessLevel(AttributeContext context, VariableNode node) throws UaException {
-                Optional<Object> identity = context.getSession().map(Session::getIdentityObject);
-
-                // no session means the write is "internal", allow it...
-                boolean userAdmin = identity.map("admin"::equals).orElse(true);
-
-                return userAdmin ?
-                    ubyte(AccessLevel.getMask(AccessLevel.CurrentRead, AccessLevel.CurrentWrite)) : ubyte(0);
+        node.setAttributeDelegate(new RestrictedAccessDelegate(identity -> {
+            if ("admin".equals(identity)) {
+                return AccessLevel.READ_WRITE;
+            } else {
+                return AccessLevel.NONE;
             }
-
-            @Override
-            protected DataValue getValue(AttributeContext context, VariableNode node) throws UaException {
-                Optional<Object> identity = context.getSession().map(Session::getIdentityObject);
-
-                // no session means the write is "internal", allow it...
-                boolean userCanRead = identity.map("admin"::equals).orElse(true);
-
-                if (userCanRead) {
-                    logger.info(
-                        "Allowing user '{}' access reading Value of {}",
-                        identity, node.getNodeId());
-
-                    return node.getValue();
-                } else {
-                    logger.info(
-                        "Denying user '{}' access reading Value of {}",
-                        identity, node.getNodeId());
-
-                    throw new UaException(StatusCodes.Bad_UserAccessDenied);
-                }
-            }
-
-            @Override
-            protected void setValue(AttributeContext context, VariableNode node, DataValue value) throws UaException {
-                Optional<Object> identity = context.getSession().map(Session::getIdentityObject);
-
-                // no session means the write is "internal", allow it...
-                boolean userCanWrite = identity.map("admin"::equals).orElse(true);
-
-                if (userCanWrite) {
-                    logger.info(
-                        "Allowing user '{}' access writing to Value of {}",
-                        identity, node.getNodeId());
-
-                    node.setValue(value);
-                } else {
-                    logger.info(
-                        "Denying user '{}' access writing to Value of {}",
-                        identity, node.getNodeId());
-
-                    throw new UaException(StatusCodes.Bad_UserAccessDenied);
-                }
-            }
-        });
+        }));
 
         server.getNodeManager().addNode(node);
         adminFolder.addOrganizes(node);
@@ -357,43 +304,15 @@ public class ExampleNamespace implements Namespace {
             .setTypeDefinition(Identifiers.BaseDataVariableType)
             .build();
 
-        node.setValue(new DataValue(new Variant("hello, admin")));
+        node.setValue(new DataValue(new Variant("admin was here")));
 
-        node.setAttributeDelegate(new VariableNodeDelegate() {
-            @Override
-            protected UByte getUserAccessLevel(AttributeContext context, VariableNode node) throws UaException {
-                Optional<Object> identity = context.getSession().map(Session::getIdentityObject);
-
-                // no session means the write is "internal", allow it...
-                boolean userCanWrite = identity.map("admin"::equals).orElse(true);
-
-                return userCanWrite ?
-                    ubyte(AccessLevel.getMask(AccessLevel.CurrentRead, AccessLevel.CurrentWrite)) :
-                    ubyte(AccessLevel.getMask(AccessLevel.CurrentRead));
+        node.setAttributeDelegate(new RestrictedAccessDelegate(identity -> {
+            if ("admin".equals(identity)) {
+                return AccessLevel.READ_WRITE;
+            } else {
+                return AccessLevel.READ_ONLY;
             }
-
-            @Override
-            protected void setValue(AttributeContext context, VariableNode node, DataValue value) throws UaException {
-                Optional<Object> identity = context.getSession().map(Session::getIdentityObject);
-
-                // no session means the write is "internal", allow it...
-                boolean userCanWrite = identity.map("admin"::equals).orElse(true);
-
-                if (userCanWrite) {
-                    logger.info(
-                        "Allowing user '{}' access writing to Value of {}",
-                        identity, node.getNodeId());
-
-                    node.setValue(value);
-                } else {
-                    logger.info(
-                        "Denying user '{}' access writing to Value of {}",
-                        identity, node.getNodeId());
-
-                    throw new UaException(StatusCodes.Bad_UserAccessDenied);
-                }
-            }
-        });
+        }));
 
         server.getNodeManager().addNode(node);
         adminFolder.addOrganizes(node);
